Move particle factory function to a new file
This is to avoid having to avoid circularity. particle_class.py
defines Particle, and particle_collections.py imports Particle to
define ParticleList. Prior to this commit, we had the factory function
in particle_class.py, where it imported ParticleList.
